more i18n, break big components into their own files
