+++ frontend/src/config/index.js
@@ -36,4 +36,4 @@ export const DEFAULT_LOCALE = process.env.TM_DEFAULT_LOCALE || 'en';
 
 
 // PROJECT CARD CONFIG
-export const PROJECTCARD_CONTRIBUTION_SHOWN_THRESHOLD = process.env.EDITS_API_URL || 5;
+export const PROJECTCARD_CONTRIBUTION_SHOWN_THRESHOLD = process.env.PROJECTCARD_CONTRIBUTION_SHOWN_THRESHOLD || 5;
