+++ frontend/src/locales/en.json
@@ -45,8 +45,9 @@
   "footer.learn": "Learn more about OpenStreetMap.",
   "project.mainSection.title": "Generic project",
   "project.card.contributorCount": "total contributors",
-  "project.card.percentMapped": "% Mapped",
-  "project.card.percentValidated": "% Validated",
+  "project.card.lastContribution": "Last contribution",
+  "project.card.percentMapped": "Mapped",
+  "project.card.percentValidated": "Validated",
   "project.card.mapperLevelBeginner": "Beginner",
   "project.card.mapperLevelIntermediate": "Intermediate",
   "project.card.mapperLevelAdvanced": "Advanced",
